src/requirements.txt
pexpect
PyYAML
poormanslogging
